Move config module outof app package.
