+++ tests/test_app.py
@@ -1,9 +1,11 @@
+import json
 import unittest
+
+import requests_mock
+
 from app.application import app
 from app.jwt import encode, decode
-from app.config import TestingConfig, Config
-import json
-import requests_mock
+from config import Config
 
 with open('tests/test_data/my_surveys.json') as json_data:
     my_surveys_data = json.load(json_data)
