+++ tests/test_surveys.py
@@ -1,8 +1,10 @@
-import unittest
-from app.application import app
-from app.config import Config, TestingConfig
 import json
+import unittest
+
 import requests_mock
+
+from app.application import app
+from config import Config
 
 with open('tests/test_data/my_surveys.json') as json_data:
     my_surveys_data = json.load(json_data)
